further file cleanup, removing commented out code, consolidate templates to components
