+++ README.md
@@ -12,7 +12,7 @@ Web application UI for the IODA project (https://ioda.caida.org)
 6. run `yarn install`
 
 ## Run 
-the code is actually two processes
+the code is two processes:
 - yarn compiles the javascript, (watch keeps it checking for updates)
     ~~~
     yarn encore dev --watch
@@ -27,14 +27,14 @@ the code is actually two processes
     php -S 127.0.0.1:8000 -t public
     ~~~
 
-Check localhost in browser
+Check localhost:8000 in browser
 
-# possible problems and error messages
+## Possible Problems and Error Messages
 ~~~
     [ErrorException]   curl_multi_setopt(): CURLPIPE_HTTP1 is no longer supported
 ~~~
 
-I fixed it by:
+Fix with:
 1. Updating the global symfony/flex using `composer global require symfony/flex ^1.5` 
 2. Removing the `vendor/symfony/flex` directory in my project.
 3. Running `composer update`.
@@ -43,4 +43,5 @@ I fixed it by:
     PHP Fatal error:  require(): Failed opening required '/.../ioda-ui/vendor/autoload.php' (include_path='.:/usr/local/Cellar/php@7.2/7.2.25/share/php@7.2/pear') in .../ioda-ui/config/bootstrap.php on line 5
 ~~~
 
-Fix it by running `composer install` in the /ioda-ui/ folder.
+Fix with:
+1. run `composer install` in the /ioda-ui/ folder.
