+++ assets/js/Ioda/components/map/Map.js
@@ -4,7 +4,6 @@ import { humanizeNumber, shadeColor, interpolateColor } from "../../utils";
 import d3 from "d3";
 
 const mapAccessToken = "pk.eyJ1Ijoid2ViZXIwMjUiLCJhIjoiY2tmNXp5bG0wMDAzaTMxbWQzcXQ1Y3k2eCJ9.NMu5bfrybATuYQ7HdYvq-g";
-const thunderForestapiKey = "f3709489dd7c411580a4610ccb5c1370";
 
 class TopoMap extends Component {
     constructor(props) {
